fix: Fixing the receipt serialization
As per

https://github.com/slockit/in3/wiki/Ethereum-Verification-and-MerkleProof#receipt-proof

if the transaction failed, the state root should be used in place of
status. We were using 0 instead.
In the block that the second test was using there are at least some
failed transactions, and thus the particia tree was computed
incorrectly.

Also we are not feeding the key properly. I put some temporary hack
there for now, but more effort and testing is needed there.